Remove duplicated Environments types.
